Rename token generator to type generator
This also fixes some logical flaws with the previous implementation.
Previously when two types hat the same token name or in the worst case
the same argument name, an identical type was created.

This adds a TypeRegistry which creates a submodule when such an
duplicate type is encountered. It also helps to resolve types when
generating code.
This solution is pretty regarding the names of modules, but at least
this works for now.